+++ Components/Errors/errorHandling.js
@@ -1,5 +1,3 @@
-import { useRecoilState } from "recoil";
-import { errorState } from "../../utils/States/states";
 import ThrowError from "./ThrowError";
 
 export default function handleError(response) {
